Track set creations in correct controller
